Stuck on set 1 challenge 6
The instructions at http://cryptopals.com/sets/1/challenges/6/
say to normalize the hamming distance by the bit length.

I focused on first cracking the key fro challenge 5, which
took some work. After much frustration I found that I could not
get the code to work if I normalize the hamming distance, but it
does successfully crack "ICE" if I do not.

However it appears that this code applied to challenge 6 does not work,
so I'm a bit stuck.